requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
pillow>=9.0.0
requests>=2.28.0
tqdm>=4.64.0
matplotlib>=3.5.0
ipywidgets>=8.0.0
accelerate>=0.20.0
scikit-learn>=1.2.0
seaborn>=0.11.0
numpy>=1.21.0
scipy>=1.9.0
pandas>=1.5.0 